Add time-limited module checks
